pyproject.toml
[project]
name = "smatsu-st-2"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "streamlit>=1.38.0",
    "pyjwt>=2.9.0",
    "streamlit-navigation-bar>=3.3.0",
    "streamlit_option_menu>=0.3.13",
    "streamlit-option-menu>=0.3.13",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/smatsu_st_2"]
